scrapers/requirements.txt
praw
requests
beautifulsoup4
psycopg2-binary
python-dotenv
selenium